pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scamrecon"
version = "0.1.0"
description = "Tools for gathering information about scam websites"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
dependencies = [
    "requests>=2.28.0",
    "dnspython>=2.1.0",
    "tqdm>=4.64.0",
    "python-dotenv>=0.20.0",
    "pandas>=1.4.0",
    "undetected-chromedriver>=3.4.0",
    "selenium>=4.1.0",
    "click>=8.1.0",
    "colorama>=0.4.4",
    "rich>=12.0.0",
    "setuptools>=75.3.2",
    "flask>=3.0.3",
    "opencv-python>=4.11.0.86",
    "pillow>=10.4.0",
    "scikit-image>=0.21.0",
]

[project.scripts]
scamrecon = "scamrecon.cli.main_refactored:cli"
scamrecon-legacy = "scamrecon.cli.main:cli"

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "mypy",
]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
